+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.0.3] - 2022-09-09
+
+### Added
+
+- `RepoUrlParsed` exposes `repo_name` which gives the name of a repo without its potential prefix.
+
+
 ## [0.0.2] - 2022-09-06
 
 ### Changed